requirements.txt
yt-dlp
pyrogram
tgcrypto
redis
rq
requests
python-telegram-bot
psycopg2-binary
